Pressing enter on the searchbar will generate the presentation
This is the biggest bug that the website had, and the solution is
not pretty. However, it gets the job done and the disaster is
confined a small exclusion zone.

Pressing the enter key will triger the same actions as clicking the
generate button since they both now call upon generate()